vision: Stop running the vision loop in simulation
100% test friendly™

It'd be great if the vision loop could run in the simulator for local
testing, but I have no idea how to do that whilst not running it in the
tests.

+++ components/vision.py
@@ -3,6 +3,7 @@ import multiprocessing.sharedctypes
 import numpy as np
 import cv2
 import cscore as cs
+import hal
 import wpilib
 
 
@@ -12,10 +13,13 @@ class Vision:
 
     def __init__(self):
         self._data_array = multiprocessing.sharedctypes.RawArray("d", [0.0])
-        self._process_run_event = multiprocessing.Event()
-        self._process_run_event.set()
-        self._process = multiprocessing.Process(target=vision_loop, args=(self._data_array, self._process_run_event))
-        self._process.start()
+
+        if not hal.isSimulation():
+            self._process_run_event = multiprocessing.Event()
+            self._process_run_event.set()
+            self._process = multiprocessing.Process(target=vision_loop, args=(self._data_array, self._process_run_event))
+            self._process.start()
+
         self.k = 0.5
         self.smoothed_x = 0.0
 
@@ -23,9 +27,10 @@ class Vision:
         wpilib.Resource._add_global_resource(self)
 
     def free(self):
-        self._process_run_event.clear()
-        self._process.join(0.1)
-        self._process.terminate()
+        if not hal.isSimulation():
+            self._process_run_event.clear()
+            self._process.join(0.1)
+            self._process.terminate()
 
     def setup(self):
         """Run just after createObjects.
